feat: bar charts stacked and colored by severity text
