+++ packages/app/src/DBSearchPage.tsx
@@ -1314,6 +1314,7 @@ function DBSearchPage() {
                               },
                             ],
                             orderBy: undefined,
+                            groupBy: searchedSource?.severityTextExpression,
                             granularity: 'auto',
                             dateRange: searchedTimeRange,
                             displayType: DisplayType.StackedBar,
