+++ packages/app/src/components/DBTimeChart.tsx
@@ -4,7 +4,6 @@ import cx from 'classnames';
 import { add } from 'date-fns';
 import {
   ClickHouseQueryError,
-  formatResponseForTimeChart,
 } from '@hyperdx/common-utils/dist/clickhouse';
 import {
   ChartConfigWithDateRange,
@@ -15,6 +14,7 @@ import { useDisclosure } from '@mantine/hooks';
 
 import {
   convertDateRangeToGranularityString,
+  formatResponseForTimeChart,
   Granularity,
   useTimeChartSettings,
 } from '@/ChartUtils';
